requirements.txt
requests>=2.25.1
pandas>=1.2.0
python-dotenv>=0.19.0
streamlit>=1.28.0
selenium>=4.0.0
webdriver-manager>=3.8.0
undetected-chromedriver>=3.5.0
